fix scene loader tests
