+++ packages/engine/src/scene/components/DirectionalLightComponent.ts
@@ -1,7 +1,8 @@
-import { Color, Vector2 } from 'three'
+import { Color, DirectionalLight, Vector2 } from 'three'
 import { createMappedComponent } from '../../ecs/functions/ComponentFunctions'
 
 export type DirectionalLightComponentType = {
+  light: DirectionalLight
   color: Color
   intensity: number
   castShadow: boolean
