+++ packages/engine/src/scene/functions/loaders/AudioFunctions.test.ts
@@ -1,0 +1,6 @@
+import assert from 'assert'
+
+describe.skip('AudioFunctions', () => {
+  it('deserializeAudio', () => {
+  })
+})
