+++ packages/engine/src/scene/functions/loaders/CameraPropertiesFunctions.test.ts
@@ -1,0 +1,6 @@
+import assert from 'assert'
+
+describe.skip('TODO Functions', () => {
+  it('deserialize TODO', () => {
+  })
+})
